Update NR_log_ingestion.py
There was an issue writing the updated file to Blob. Two rows commented out so that the file stayed in the C: and could be uploaded manually to Blob. The two rows commented out effectively cut the document from C: and paste to Blob so when the code finished running, there was no back up copy of the updated file.